fix updateEditorForFileUrl
